# lib/api-zod/src/generated/types/createLessonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Control API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLessonRequestLessonType } from "./createLessonRequestLessonType";

export interface CreateLessonRequest {
  date: string;
  topic: string;
  lessonType: CreateLessonRequestLessonType;
  groupId: number;
  subjectId: number;
}