# lib/api-zod/src/generated/types/createStudentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Control API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStudentRequest {
  firstName: string;
  lastName: string;
  middleName?: string;
  groupId: number;
  studentNumber?: string;
}