# lib/api-zod/src/generated/types/groupReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Control API
 * OpenAPI spec version: 0.1.0
 */
import type { Group } from "./group";
import type { GroupReportAttendanceMatrixItem } from "./groupReportAttendanceMatrixItem";
import type { Lesson } from "./lesson";
import type { StudentAttendanceSummary } from "./studentAttendanceSummary";
import type { Subject } from "./subject";

export interface GroupReport {
  group: Group;
  subject?: Subject;
  lessons: Lesson[];
  students: StudentAttendanceSummary[];
  attendanceMatrix: GroupReportAttendanceMatrixItem[];
}